Sp 20180122 Engine Update

+++ OpenRA.Mods.Common/AI/AttackOrFleeFuzzy.cs
@@ -1,6 +1,6 @@
 #region Copyright & License Information
 /*
- * Copyright 2007-2017 The OpenRA Developers (see AUTHORS)
+ * Copyright 2007-2018 The OpenRA Developers (see AUTHORS)
  * This file is part of OpenRA, which is free software. It is made
  * available to you under the terms of the GNU General Public License
  * as published by the Free Software Foundation, either version 3 of
@@ -198,7 +198,8 @@ namespace OpenRA.Mods.Common.AI
 			if (sumOfMaxHp == 0)
 				return 0.0f;
 
-			return (sumOfHp * normalizeByValue) / sumOfMaxHp;
+			// Cast to long to avoid overflow when multiplying by the health
+			return (int)((long)sumOfHp * normalizeByValue / sumOfMaxHp);
 		}
 
 		static float RelativePower(IEnumerable<Actor> own, IEnumerable<Actor> enemy)
