+++ OpenRA.Mods.Common/Activities/Transform.cs
@@ -1,6 +1,6 @@
 #region Copyright & License Information
 /*
- * Copyright 2007-2017 The OpenRA Developers (see AUTHORS)
+ * Copyright 2007-2018 The OpenRA Developers (see AUTHORS)
  * This file is part of OpenRA, which is free software. It is made
  * available to you under the terms of the GNU General Public License
  * as published by the Free Software Foundation, either version 3 of
@@ -122,7 +122,8 @@ namespace OpenRA.Mods.Common.Activities
 				var health = self.TraitOrDefault<Health>();
 				if (health != null)
 				{
-					var newHP = ForceHealthPercentage > 0 ? ForceHealthPercentage : (health.HP * 100) / health.MaxHP;
+					// Cast to long to avoid overflow when multiplying by the health
+					var newHP = ForceHealthPercentage > 0 ? ForceHealthPercentage : (int)(health.HP * 100L / health.MaxHP);
 					init.Add(new HealthInit(newHP));
 				}
