+++ OpenRA.Mods.Common/Traits/DamagedByTerrain.cs
@@ -1,6 +1,6 @@
 #region Copyright & License Information
 /*
- * Copyright 2007-2017 The OpenRA Developers (see AUTHORS)
+ * Copyright 2007-2018 The OpenRA Developers (see AUTHORS)
  * This file is part of OpenRA, which is free software. It is made
  * available to you under the terms of the GNU General Public License
  * as published by the Free Software Foundation, either version 3 of
@@ -69,7 +69,8 @@ namespace OpenRA.Mods.Common.Traits
 			if (totalTiles == 0)
 				return;
 
-			damageThreshold = (Info.DamageThreshold * health.MaxHP + (100 - Info.DamageThreshold) * safeTiles * health.MaxHP / totalTiles) / 100;
+			// Cast to long to avoid overflow when multiplying by the health
+			damageThreshold = (int)((Info.DamageThreshold * (long)health.MaxHP + (100 - Info.DamageThreshold) * safeTiles * (long)health.MaxHP / totalTiles) / 100);
 
 			// Actors start with maximum damage applied
 			var delta = health.HP - damageThreshold;
