+++ OpenRA.Mods.Common/Traits/Explodes.cs
@@ -1,6 +1,6 @@
 #region Copyright & License Information
 /*
- * Copyright 2007-2017 The OpenRA Developers (see AUTHORS)
+ * Copyright 2007-2018 The OpenRA Developers (see AUTHORS)
  * This file is part of OpenRA, which is free software. It is made
  * available to you under the terms of the GNU General Public License
  * as published by the Free Software Foundation, either version 3 of
@@ -134,7 +134,8 @@ namespace OpenRA.Mods.Common.Traits
 			if (Info.DamageThreshold == 0)
 				return;
 
-			if (health.HP * 100 < Info.DamageThreshold * health.MaxHP)
+			// Cast to long to avoid overflow when multiplying by the health
+			if (health.HP * 100L < Info.DamageThreshold * (long)health.MaxHP)
 				self.World.AddFrameEndTask(w => self.Kill(e.Attacker));
 		}
 	}
