+++ OpenRA.Mods.Common/Traits/Health.cs
@@ -1,6 +1,6 @@
 #region Copyright & License Information
 /*
- * Copyright 2007-2017 The OpenRA Developers (see AUTHORS)
+ * Copyright 2007-2018 The OpenRA Developers (see AUTHORS)
  * This file is part of OpenRA, which is free software. It is made
  * available to you under the terms of the GNU General Public License
  * as published by the Free Software Foundation, either version 3 of
@@ -43,7 +43,8 @@ namespace OpenRA.Mods.Common.Traits
 			Info = info;
 			MaxHP = info.HP > 0 ? info.HP : 1;
 
-			hp = init.Contains<HealthInit>() ? init.Get<HealthInit, int>() * MaxHP / 100 : MaxHP;
+			// Cast to long to avoid overflow when multiplying by the health
+			hp = init.Contains<HealthInit>() ? (int)(init.Get<HealthInit, int>() * (long)MaxHP / 100) : MaxHP;
 
 			DisplayHP = hp;
 		}
